Prettyprinter.Util

Description

Frequently useful definitions for working with general prettyprinters.

Documentation

words :: Text -> [Doc ann]

Split an input into word-sized Docs.

>>> putDoc (tupled (words "Lorem ipsum dolor"))
(Lorem, ipsum, dolor)

putDocW :: Int -> Doc ann -> IO ()

Render a document with a certain width. Useful for quick-and-dirty testing of layout behaviour. Used heavily in the doctests of this package, for example.

>>> let doc = reflow "Lorem ipsum dolor sit amet, consectetur adipisicing elit"
>>> putDocW 20 doc
Lorem ipsum dolor
sit amet,
consectetur
adipisicing elit
>>> putDocW 30 doc
Lorem ipsum dolor sit amet,
consectetur adipisicing elit

reflow :: Text -> Doc ann

Insert soft linebreaks between words, so that text is broken into multiple lines when it exceeds the available width.

>>> putDocW 32 (reflow "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.")
Lorem ipsum dolor sit amet,
consectetur adipisicing elit,
sed do eiusmod tempor incididunt
ut labore et dolore magna
aliqua.

reflow = fillSep . words