Data.Text.Prettyprint.Doc.Render.Util.Panic

Documentation

panicUncaughtFail :: void

Raise a hard error if there is a SFail in a SimpleDocStream.

panicUnpairedPop :: void

Raise a hard error when an annotation terminator is encountered in an unannotated region.

panicSimpleDocTreeConversionFailed :: void

Raise a hard generic error when the SimpleDocStream to SimpleDocTree conversion fails.

panicInputNotFullyConsumed :: void

Raise a hard error when the »to SimpleDocTree« parser finishes without consuming the full input.

panicPeekedEmpty :: void

panicPoppedEmpty :: void